refactor: merge CommonWebContentsDelegate into api::WebContents (#26189)
