+++ filenames.gni
@@ -127,7 +127,6 @@ filenames = {
     "shell/browser/api/electron_api_web_contents_mac.mm",
     "shell/browser/auto_updater_mac.mm",
     "shell/browser/browser_mac.mm",
-    "shell/browser/common_web_contents_delegate_mac.mm",
     "shell/browser/electron_browser_main_parts_mac.mm",
     "shell/browser/mac/dict_util.h",
     "shell/browser/mac/dict_util.mm",
@@ -203,7 +202,6 @@ filenames = {
     "shell/browser/api/electron_api_browser_window_views.cc",
     "shell/browser/api/electron_api_menu_views.cc",
     "shell/browser/api/electron_api_menu_views.h",
-    "shell/browser/common_web_contents_delegate_views.cc",
     "shell/browser/native_browser_view_views.cc",
     "shell/browser/native_browser_view_views.h",
     "shell/browser/native_window_views.cc",
@@ -336,8 +334,6 @@ filenames = {
     "shell/browser/browser_process_impl.h",
     "shell/browser/child_web_contents_tracker.cc",
     "shell/browser/child_web_contents_tracker.h",
-    "shell/browser/common_web_contents_delegate.cc",
-    "shell/browser/common_web_contents_delegate.h",
     "shell/browser/cookie_change_notifier.cc",
     "shell/browser/cookie_change_notifier.h",
     "shell/browser/electron_autofill_driver.cc",
